pyproject.toml
[tool.poetry]
name = "AliceBob"
version = "0.1.0"
description = ""
authors = ["Alice & Bob Yeam <hello@alicebob.io>"]
packages = [{ include = "alicebob"}, { include = "ezycourse"}]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
notion-client = "^2.2.1"
lxml = "^5.3.0"
requests = "^2.32.3"
anthropic = "^0.39.0"
html2text-rs = "^0.2.0"
celery = {extras = ["django", "librabbitmq", "redis"], version = "^5.4.0"}
gevent = "^24.11.1"
gunicorn = "^23.0.0"
python-decouple = "^3.8"
orjson = "^3.10.11"
ollama = "^0.3.3"
sentry-sdk = {extras = ["celery"], version = "^2.18.0"}
django = {extras = ["argon2"], version = "^5.1.3"}
click = "^8.1.7"
rich = "^13.9.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
